requirements.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.0.7
django-allauth==0.36.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.0.1
django-mdeditor==0.1.9
django-pure-pagination==0.3.0
django-reversion==2.0.13
django-suit==0.2.26
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.11.3
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.5.4
Pillow==5.2.0
pycparser==2.18
PyJWT==1.6.4
PyMySQL==0.9.2
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-core==1.7.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
xlrd==1.1.0
xlwt==1.3.0
